+++ CHANGELOG.md
@@ -15,6 +15,10 @@ All notable changes to this project are documented in this file.
 
 ## upcomming
 
+- [BREAKING] dropping the LayerId does not cause the layer to be removed. Only if the layer is also empty does that happen now.
+- [BREAKING] LayerId is now !Send.
+- Implement clone for LayerId.
+
 ## 0.1.0-alpha0.4
 
 - Improve the `all` example